# docs/MES-CoBraD_ES-OpenAPI.yml
openapi: 3.0.1
info:
  title: Swagger API for ES
  description: This is an API structure for the ES interoperability.
  termsOfService: http://swagger.io/terms/
  contact:
    email: johnmtsi@gmial.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://es.mes-cobrad.eu
...
paths:
  /workflow:
    get:
      tags:
        - Workflow
      summary: Get all workflows
      operationId: getWorkflow
      responses:
        200:
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
        400:
          description: Invalid workflow value
          content: {}
    post:
      tags:
        - Workflow
      summary: Post a new workflow
      operationId: addWorkflow
      requestBody:
        description: Workflow object that needs to be added to the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Workflow"
          application/xml:
            schema:
              $ref: "#/components/schemas/Workflow"
        required: true
      responses:
        default:
          description: successfully created
          content: {}
      x-codegen-request-body-name: body
  /workflow/{workflowid}:
    get:
      tags:
        - Workflow
      summary: Get a workflow with a specific Id
      operationId: getWorkflowId
      parameters:
        - name: workflowid
          in: path
          description: Workflow ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/Workflow"
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: workflow not found
          content: {}
    put:
      tags:
        - Workflow
      summary: Put a workflow with a specific Id
      operationId: putWorkflowId
      parameters:
        - name: workflowid
          in: path
          description: Workflow ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Update workflows
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/Workflow"
        required: true
      responses:
        200:
          description: OK
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - Workflow
      summary: Delete workflow
      description: This can only be done by the logged in user.
      operationId: deleteWorkflow
      parameters:
        - name: workflowid
          in: path
          description: The Id that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        400:
          description: Invalid Id supplied
          content: {}
        404:
          description: Id not found
          content: {}
  /workflow/{workflowid}/run:
    get:
      tags:
        - Workflow
      summary: Runs a workflow with Id
      operationId: runWorkflow
      parameters:
        - name: workflowid
          in: path
          description: Id of workflow to run
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        404:
          description: Id not found
          content: {}
...
components:
  schemas:
    Workflow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: workflow-therapeutic1
        category:
          $ref: "#/components/schemas/Category"
        run:
          $ref: "#/components/schemas/Run"
        variable:
          $ref: "#/components/schemas/Variable"
        step:
          $ref: "#/components/schemas/Step"
        metadata:
          $ref: "#/components/schemas/Metadata"
        status:
          type: string
          description: workflow status in the general process
          enum:
            - completed
            - uncomplete
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
    Step:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Run:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: boolean
    Variable:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Metadata:
      type: object
      properties:
        id:
          type: string
          format: uuid
        data:
          type: object
      description: Metadata for further use
...